x509: split package apart, try to clean up API
